feat (tests): Adding initial set of tests for REST
Related to #357
