+++ primeqa/services/constants.py
@@ -2,7 +2,8 @@ from enum import Enum
 
 ATTR_INDEX_ID = "index_id"
 ATTR_STATUS = "status"
-ATTR_ENGINE_TYPE  ="engine_type"
+ATTR_ENGINE_TYPE = "engine_type"
+ATTR_METADATA = "metadata"
 
 
 class IndexStatus(str, Enum):
