+++ primeqa/services/factories.py
@@ -10,6 +10,7 @@ from primeqa.components.base import (
     Indexer,
 )
 from primeqa.components.reader.extractive import ExtractiveReader
+from primeqa.components.reader.generative import GenerativeFiDReader
 
 from primeqa.components.retriever.dense import ColBERTRetriever, DPRRetriever
 from primeqa.components.retriever.sparse import BM25Retriever
@@ -20,6 +21,7 @@ from primeqa.components.indexer.sparse import BM25Indexer
 
 READERS_REGISTRY = {
     ExtractiveReader.__name__: ExtractiveReader,
+    GenerativeFiDReader.__name__: GenerativeFiDReader,
 }
 
 RETRIEVERS_REGISTRY = {
