+++ tests/primeqa/services/rest_server/test_readers.py
@@ -1,0 +1,28 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+#
+# Copyright 2022-2023 PrimeQA Team
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+def test_get_readers(mock_client):
+    response = mock_client.get(
+        "/readers",
+    )
+    assert response.status_code == 200
+    readers = response.json()
+    assert len(readers) == 2
+    assert ["ExtractiveReader", "GenerativeFiDReader"] == [
+        reader["reader_id"] for reader in readers
+    ]
